falcon-containment-automation/requirements.txt
falconpy>=1.4.9
python-dotenv>=1.0.1
